why wasnt this like this to begin with!

+++ include/bit_field.h
@@ -257,7 +257,7 @@ static inline size_t get_bit_field_size(
 //! \param[in] b The sequence of words representing a bit_field.
 //! \param[in] s The size of the bit_field.
 //! \return The function returns the number of bits set to 1.
-static inline int count_bit_field(
+static inline uint32_t count_bit_field(
 	bit_field_t b,
 	size_t s)
 {
@@ -266,7 +266,7 @@ static inline int count_bit_field(
     for ( ; s > 0; s--) {
 	sum += __builtin_popcount(b[s - 1]);
     }
-    return sum;
+    return (uint32_t) sum;
 }
 
 //! \brief Prints a bit_field as ones and zeros.
